Refactored DeleteAllBooksServiceImpl

+++ v1/src/main/java/com/api/v1/book/services/DeleteAllBooksServiceImpl.java
@@ -1,6 +1,7 @@
 package com.api.v1.book.services;
 
 import com.api.v1.book.domain.BookRepository;
+import com.api.v1.book.exceptions.BookEntityNotExistException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
@@ -13,7 +14,13 @@ class DeleteAllBooksServiceImpl implements DeleteAllBooksService {
 
     @Override
     public Mono<Void> deleteAll() {
-        return repository.deleteAll();
+        return repository
+                .findAll()
+                .hasElements()
+                .flatMap(exists -> {
+                    if (!exists) return Mono.error(BookEntityNotExistException::new);
+                    return repository.deleteAll();
+                });
     }
 
 }
